rust: Split openat helpers into module
Will help use them in other parts of our code.
