+++ rust/src/composepost.rs
@@ -19,33 +19,11 @@
 use failure::Fallible;
 use openat;
 use rayon::prelude::*;
+use std::io;
 use std::io::{BufRead, Write};
 use std::path::Path;
-use std::{fs, io};
 
-/// Helper functions for openat::Dir
-trait OpenatDirExt {
-    // IMO should propose this at least in a "utils" bit of the openat crate;
-    // Like 95% of the time I'm looking at errno (with files) it's for ENOENT,
-    // and Rust has an elegant way to map that with Option<>.  Every other
-    // error I usually just want to propagate back up.
-    fn open_file_optional<P: openat::AsPath>(&self, p: P) -> io::Result<Option<fs::File>>;
-}
-
-impl OpenatDirExt for openat::Dir {
-    fn open_file_optional<P: openat::AsPath>(&self, p: P) -> io::Result<Option<fs::File>> {
-        match self.open_file(p) {
-            Ok(f) => Ok(Some(f)),
-            Err(e) => {
-                if e.kind() == io::ErrorKind::NotFound {
-                    Ok(None)
-                } else {
-                    Err(e)
-                }
-            }
-        }
-    }
-}
+use openat_utils::OpenatDirExt;
 
 // rpm-ostree uses /home → /var/home by default as generated by our
 // rootfs; we don't expect people to change this.  Let's be nice
