+++ rust/src/lib.rs
@@ -49,3 +49,4 @@ mod journal;
 pub use journal::*;
 mod utils;
 pub use utils::*;
+mod openat_utils;
